Periodically send status to systemd (#3006)
* Periodically send status to systemd

With systemd it's possible to send a status line. This will show up when
users run systemctl status puma.service.

Most of this code is based on puma-plugin-systemd.

* Rewrite systemd integration as a plugin

The primary motiviation for this is that plugins have native integration
for background threads. This is much cleaner since it allows tracking of
those. For example, it's possible to clean up those threads in the test
suite.

* add test for systemd plugin and tweak message fetch logic

---------

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>