Fix warn_if_in_single_mode incorrect message (#3111)
The message could either be in the positive `but Puma is configured to run in single mode (worker count = 0 )` or the negative `but Puma is not configured to run in cluster mode (worker count > 0 )`. I went with negative because it explicitly instructs what is required in order for hooks to run.