Make puma cluster process suitable as PID 1
We recently ran into an issue in production where our puma
containers had thousands of zombie (defunct) processes.

This was caused by the web application spawning some multi-process
command (Google Chrome) and sometimes interrupting it hard.

- pid=1 puma cluster
 \- pid=2 puma worker
  \- pid=3 google chrome
   \- pid=4 google chrome subprocess

In such scenario if pid=3 dies without first reaping pid=4,
then pid=4 get reparented as a child of pid=1, and pid=1 is
responsible to reap it.

On classic style hosting, PID 1 is the init system (e.g. systemd),
but with containers PID 1 is the container initial command, in our
case the Puma cluster process.

Some people uses some minimal init implementations like `tini` to
handle this, but I believe this enter the category of "unknown unknowns",
as it you need to know about this concern to avoid the problem.

This commit simply change the `wait_workers` method to reap all childrens
not just the workers it knows about.

If it end up reaping an unknown process, it logs it.

This is pretty much exactly what Unicorn and Pitchfork do, and remove
the need for an init system in containers.