Return 405, not 501, when an unsupported HTTP method is used
Currently, when the `supported_http_methods` DSL config option is not
set to `:any`, the behaviour upon encountering an unexpected HTTP method
is to return an `501 Not Implemented` response.

Whilst this is technically correct, it can be somewhat problematic; in
situations where 5xx error responses are used as a heuristic to
determine whether or not an application is healthy, a single person
sending a flood of requests with invalid HTTP methods can result in an
application being considered 'unhealthy' because Puma returns a 501.

It seems the somewhat more semantically correct response code to use in
this scenario is `405 Method Not Allowed` – this is how Rails deals with
HTTP methods it can't handle, for example. The HTTP spec requires that
an `Allow` header is returned alongside it with details of which methods
_are_ acceptable.

To that end, this modifies the behaviour of Puma to return a 405 when
supplied an unsupported HTTP method, rather than a 501, and return an
`Allow` header with the valid methods.