Allow to configure both min and max threads as a single number
I may be off, but it seems to be that in the overwhelming
majority of use cases, one will want a fixed number of threads
rather than a mix/max, and the `--threads` CLI argument
already allow to pass a single number.

I such I think it would simplify many configs.