Print RUBY_DESCRIPTION when Puma starts
This commit changes the runner to print `RUBY_DESCRIPTION` instead of
a manually-generated string for the Ruby version. Other popular gems
like Sidekiq similarly print `RUBY_DESCRIPTION`. One specific motivation
for this change is that `RUBY_DESCRIPTION` indicates whether YJIT is
enabled in MRI.

This commit also deprecates the `Puma::Runner#ruby_engine` method,
which will be removed in v7.