lib/client.rb - use Tempfile.create instead of Tempfile.new for request bodies (#3296)
Currently, `Tempfile.new` is used in `Client` when a request contains content (a body) that is either very large or indeterminate (chunked).  `Tempfile.new` creates a finalizer that removes the `Tempfile`, so its removal is dependent on GC.

Also, in general, I think we'd prefer that Puma does its own cleanup, and not depend on GC to do so, at least with system resources like IO (files, pipes, sockets, etc).

So, use `Tempfile.create` and remove the file when Puma is finished with it.