[CI] test_integration_cluster.rb - remove blocking `@server.gets` (#3471)
Naked `@server.gets` calls are blocking.  Parallel testing will always be unstable (and unpredictable) with blocking calls.

But, due to the resources used by runner VM's and not used by 'local' testing, a test file that runs fine locally may be unstable when run on a multi VM CI system.

Hence, remove all `@server.gets` calls and replace with helper methods that are not blocking.