[CI] test_launcher.rb - fixup test_puma_stats_clustered, small refactor (#3499)
While running updates to the test suite, this test failed in two macOS jobs. I then noticed that it reliably failed on my desktop.

test_launcher.rb does not create client requests, so there are no changes to it in the new test framework.

Use a queue to let the launcher stop. Seems to fix the issues with thread contention...