[CI] test_puma_server.rb - add HTTP/1.0 & HTTP/1.1 checks thru to app env (#3525)
The two tests are simple tests that checked `env['SERVER_PROTOCOL']` for both HTTP/1.0 & HTTP/1.1.  They checked thru to the values passed to the app.

Similar tests also exist in the `test_http10.rb` and `test_http10.rb`, but those check the return from the parser.  Not sure if these are needed, but they don't take long to run.