Add initial Puma support document
The goal is to encode what we're doing today into something an external observer can reference. This is NOT to be used to overrule a maintainer's future decisions or to take power out of their hands. This should eventually become a set of "sensible defaults" that make it easier for contributors and users to understand desired behavior and our general roadmap.

I'm interested in specific feedback about details listed here (please make it as an inline code suggestion if possible). I'm also interested in whether there are other external things we're relying on (rack spec is external, ruby version is external, etc.) that should be added to this document.