Remove Ruby <= 2.4 bugfix/workaround
Patch https://github.com/puma/puma/pull/1345 was added as a workaround for https://bugs.ruby-lang.org/issues/13632. 

The gem addition was already removed in another commit. An un-monkey-patched version of Ruby does not have this method, so the respond_to? check always returns false, so this code is currently a no-op anyway.

```
(irb):2:in '<main>': undefined method 'purge_interrupt_queue' for an instance of Thread (NoMethodError)
	from <internal:kernel>:168:in 'Kernel#loop'
	from /Users/rschneeman/.gem/ruby/3.4.5/gems/irb-1.15.2/exe/irb:9:in '<top (required)>'
	from /Users/rschneeman/.gem/ruby/3.4.5/bin/irb:25:in 'Kernel#load'
	from /Users/rschneeman/.gem/ruby/3.4.5/bin/irb:25:in '<main>'
```