[Note this is stacked on top of 6.6.1] - it should cherry-pick on top of `4e60f73` which is puma-7.0.0.pre1 (also tagged at https://github.com/puma/puma/releases/tag/v7.0.0.pre1).
This is intended to help debugging #3715 the idea is that we could modify puma to output profiling data to understand where time is being spent. I'm unsure if this is the correct approach. But it shows AN idea of how to profile what's going on inside of puma (instead of just guessing). Long term it would be cool to make this flexible and able to be toggleable similar to https://github.com/ruby/syntax_suggest/blob/8c13ff82656514f9481304da9b9d0bbd6923cceb/spec/spec_helper.rb#L72, but for now let's hardcode everything.

Ruby-prof is a massive shotgun for performance work, i.e. don't ever use it in production. It WILL slow your program to a crawl. But the data can be very valuable if you skim performance PRs on syntax_suggest you'll see I used it to great effect, for example https://github.com/ruby/syntax_suggest/pull/104. (though normally I just look at the HTML output and scan for outliers). It's much harder here because it will capture not just what's happening in our codebase, but anything in rack and the app which SHOULD take up the bulk of time (in an ideal world). 

You can use it RIGHT NOW:

```
$ PUMA_RUBY_PROF_DIR=tmp bundle exec ruby -w 1 -Ilib bin/puma -w 1 test/rackup/hello.ru
```

When the server exits it will write profiling data to `tmp/<timestamp>` along with a symlink `tmp/last` to the most recent profile run. Then inside there will be N directories for N workers each with separate profile data.

Another thing to be mindful about is that profiling is "on" even when the server is doing no work. That means, any time spent before you're getting a request and any time spent after requests are done, but you've not killed the server yet, still count against you. That might skew our results i.e. if we spend 5 s before sending data, then send data for 5s and it takes us 5s to kill the server ... it might look like it spent FOREVER waiting to accept a new connection, but that is really just the server idling. I'm not sure how to account for that skew, just be mindful it's there. Suggestions welcome.

It also wouldn't catch things like "one worker is being greedy with accept calls and starving others" thankfully, the perf bug was reproducible with threads=1 workers=1 so we don't have to care about weird distributed emergent behavior.

Also, no need to ONLY use ruby-prof, you can use any profiler you want. If you feel like another one might be good (I found a lot of un-needed allocations with memory profiler in Rails) or there's vernier and some other sampling profilers. i.e. don't assume this is THE profiler, it's just one I've had success with in the past.

I'm thinking that the next step would be to validate this output is useful in isolation and/or modify the TechEmpower benchmark to:

- Use ruby-prof-d puma code
- Make the output of profile folder show up on the local machine (the tests run in a docker container, need to set mount a local directory and set PUMA_RUBY_PROF_DIR to the correct location).

Then once that's set:

- Profile, try to find where things are slow, make changes
- Re run TechEmpower with the attempted fix (and without ruby-prof) and see if it improves the results of Puma 7.0.0.pre1 versus Puma 6.6.1

This is overly verbose because: I'm pretty busy with my family this weekend, and I'm flying to Amsterdam on tuesday (monday is a holiday). I'm hoping people can take this suggestion and run with it if they have time. I'll try to poke at the problem on the plane but ... limited wifi etc. Don't assume I'll figure this out, many eyes make short work of bugs. The more hacking at this the faster it will go.