puma_http11.c: Use interned UTF-8 strings for hash keys
`rb_str_new2` creates a mutable, ASCII-8BIT (binary) strings.

Hence, when used with `rb_hash_aset`, ruby always has to lookup
the equivalent string in the fstring table, likely fail to find it,
then dup and free the given string, and use that as the hash key.

By keeping interned strings in these global variables, we can save
Ruby quite a bit of work. It's not a whole lot in the grand scheme
of things, but still welcome.