Move sleep cluster logic to its own class (#3746)
* Move cluster logic to its own class

Simplify the calculation. This is an alternative to #3723.

Puma 6.6.1 behavior documented here https://gist.github.com/schneems/cef38e9448dfb72943d13050a7da0869

This changes the puma 7 behavior to:

- Starts sleeping the event loop before being overloaded (the prior `pool.busy_threads` included in the ready @todo queue) and would only fire. This is closer to puma 6.6.1 behavior
- Sleeps a consistent proportional amount, closer to puma 7 behavior, versus puma 6 used a static value.
- Sleeping behavior is restricted to 2 or more workers, verus puma 7 was accidentally enabled for clusters with only 1 worker

Fixes #3740. This `wait_for_less_busy_worker` value is now treated as a maximum value for the sleep calculation. It also now respects a value of 0 to mean "don't sleep at all" which was the prior behavior of that value.

* Whitespace

* Remove unused method

The busy_thread call is now accessed directly through the thread pool and only used in one place.

* Add docs about server <-> reactor relationship

The code doesn't make it clear who is calling whom when looking at it in isolation.

* Make maximum 25x thread count and allow for overage

Puma can take in more requests than it has threads. This change preserves the same logic as before, but it doesn't hit maximum sleep until it is at 25x the number of max threads. That means that if a server was using 5 threads,  before it would hit 0.005 sleep if those five threads were busy, now if (only) five threads are busy it will sleep 0.0002 seconds.

* Move order of comparison

If busy threads is zero we don't need to calculate percentage.

* Refactor out clamp

Because we're clamped at the numerator, the division will naturally tend towards 1.0. We don't need a second clamp on the result. This allows us to remove an intermediate variable and multiply directly on the return calculation.

* Lazy max_threads

There are proposals to make thread counts dynamic RE https://github.com/puma/puma/pull/3658. If this happens we need to pull in the current value, and cannot rely on it being set once.

* Avoid re-checking the `max_delay` number every iteration

* Refactor out branching

The equation already holds up the tested properties without needing to special-case when busy threads is zero. Removes a branching conditional.

* Push worker logic into delay class

The logic of whether or not the calculation should be performed based on worker count can be calculated once and not repeated on every iteration.

* Optimize case where no threads are busy

Moves the condition from checking if the delay is zero to checking if the input is zero earlier. This is prevents some calculations and preserves the same number of conditionals.