Fix SSL_shutdown error handling (#3703)
* Fix SSL_shutdown error handling

SSL_shutdown() call may result in error and in OpenSSL every error
should be consumed by the caller via SSL_get_error() or a similar func.

MiniSSL implementation was breaking this convention so it resulted
in cryptic bugs like SSL clients (not related to Puma in any way)
return errors with text like "SSL_read: shutdown while in init" that
is initially generated by the Puma reactor but not consumed by it.

Such bugs are really hard to debug...

* Add TestPumaServerSSL#test_http_1*_close_no_errors