Remove global side effect from test (#3765)
## Before

This test asserts that when certain configuration options are specified that a custom load error is raised if concurrent-ruby cannot be found. It does this by: 

- finding the real file on disk-
- renaming the file
- running code and asserting an error with a specific message
- ensuring the file rename is reverted

Most developers would be surprised that running the test suite of one project might cause other code running on the same machine to fail, but because these modifications to disk are global, that's what happens.

## After

Instead of making a real system change, this mocks the interface where that file is loaded i.e. the "require" call:

```ruby
    def require_processor_counter
      require 'concurrent/utility/processor_counter'
    rescue LoadError
```

This is done by monkey patching the `require` method on the specific instance of Configuration under test. This means that not only does it have no side effects that are visible by other running processes on the same machine, other instances of the configuration class won't be affected.

To get this strategy to work we have to force the instance to re-parse the environment by explicitly calling the `puma_default_options` method (which in turn eventually calls the code that generates the warning). This is needed because the Puma::Configuration class eagerly evaluates the environment and calls this logic in the constructor

```ruby
module Puma
  # ...
  class Configuration
    # ...
    def initialize(user_options={}, default_options = {}, env = ENV, &block)
      default_options = self.puma_default_options(env).merge(default_options)
```