Fix Flappy test: [test/test_thread_pool.rb:362]
Analysis is here https://github.com/puma/puma/issues/3771#issuecomment-3324625876. I changed this to a queue and was able to run for many minutes without seeing a failure

```
$ while be m test/test_thread_pool.rb:339; do echo "Test passed, running again..."; done
```

So I think the cause is the `<<` call not being atomic in JRuby versus with MRI it is (due to the GIL). The fix is to switch to using a queue, but this does not solve the other observed failure

```
TestThreadPool#test_shutdown_with_grace [test/test_thread_pool.rb:369]:
Expected: 2
```