Fix flappy test: test/test_thread_pool.rb:178
Before:

```
$ while be m test/test_thread_pool.rb:178; do echo "Test passed, running again..."; done

# ...

  1) Failure:
TestThreadPool#test_trim [test/test_thread_pool.rb:178]:
Expected: 0
  Actual: 1
```

After: Runs for several minutes with no failures.

```
$ while be m test/test_thread_pool.rb:187; do echo "Test passed, running again..."; done
```

## Why

Previously we weren't waiting at all after the trim and there's definetly a race condition there. Even with the wait, there's a race condition: The @trim_requested is decremented before the @spawned is decremented. We could fix it by changing the ordering of the variable decrement but that's fragile. It's better to guarantee data safety via a mutex.