Fix flappy `test/test_thread_pool.rb` tests  (#3772)
* Fix Flappy test: [test/test_thread_pool.rb:362]

Analysis is here https://github.com/puma/puma/issues/3771#issuecomment-3324625876. I changed this to a queue and was able to run for many minutes without seeing a failure

```
$ while be m test/test_thread_pool.rb:339; do echo "Test passed, running again..."; done
```

So I think the cause is the `<<` call not being atomic in JRuby versus with MRI it is (due to the GIL). The fix is to switch to using a queue, but this does not solve the other observed failure

```
TestThreadPool#test_shutdown_with_grace [test/test_thread_pool.rb:369]:
Expected: 2
```

* Fix flappy test: test/test_thread_pool.rb:178

Before:

```
$ while be m test/test_thread_pool.rb:178; do echo "Test passed, running again..."; done

# ...

  1) Failure:
TestThreadPool#test_trim [test/test_thread_pool.rb:178]:
Expected: 0
  Actual: 1
```

After: Runs for several minutes with no failures.

```
$ while be m test/test_thread_pool.rb:187; do echo "Test passed, running again..."; done
```

## Why

Previously we weren't waiting at all after the trim and there's definetly a race condition there. Even with the wait, there's a race condition: The @trim_requested is decremented before the @spawned is decremented. We could fix it by changing the ordering of the variable decrement but that's fragile. It's better to guarantee data safety via a mutex.

* Fix flappy tests

```
$ while be m test/test_thread_pool.rb:199; do echo "Test passed, running again..."; done
```

* Fix mutex re-lock error

```
  1) Error:
TestThreadPool#test_trim_is_ignored_if_no_waiting_threads:
ThreadError: Mutex relocking by same thread
    org/jruby/ext/thread/Mutex.java:102:in `lock'
    org/jruby/ext/thread/Mutex.java:181:in `synchronize'
    test/test_thread_pool.rb:62:in `block in trim'
    org/jruby/RubyKernel.java:1725:in `loop'
    test/test_thread_pool.rb:58:in `trim'
    test/test_thread_pool.rb:222:in `block in test_trim_is_ignored_if_no_waiting_threads'
    test/test_thread_pool.rb:44:in `block in <<'
    org/jruby/ext/thread/Mutex.java:183:in `synchronize'
    lib/puma/thread_pool.rb:246:in `with_mutex'
    test/test_thread_pool.rb:42:in `<<'
    test/test_thread_pool.rb:220:in `test_trim_is_ignored_if_no_waiting_threads'
```

* Don't pass thread while in a mutex