Reintroduce keepalive "fast inline" optimization
Short: This reintroduces a performance optimization dropped in Puma 7, and adds a guard to prevent "unfair" or "longtail" request distribution when the server is overloaded.

## Context

Puma 6 and before had a "fast inline" loop. A connection can have many requests, as an optimization, this "fast inline" feature would allow a Puma thread to respond to multiple requests on the same connection in a row without requiring that the connection be placed in the queue or the thread needing to go through mutex lock/unlock-ing to get it back. This improved preformance but created a "longtail" situation where one keepalive connection can "monopolize" a thread for up to 10 requests in a row. If all threads are busy, this blocks an incoming request for 10x the average time to serve a request, this was reported in https://github.com/puma/puma/issues/3487. 

Puma 7 changed this behavior to ensure that all responses are "fair" and that a keepalive connections cannot "monopolize" a thread. This eliminated the "longtail" behavior. 

Recently https://github.com/puma/puma/issues/3788 was opened to highlight that while MRI numbers on TechEmpower benchmarks look pretty decent, the JRuby ones fell off a cliff. Previously we worked on tuning/improving the MRI performance on a similar report against 7.0.0.pre1 https://github.com/puma/puma/issues/3715. We fixed that by removing an env var from the TechEmpower setup and increasing the default keepalive connection default. This new report with JRuby seems directly related to the overhead involved in coordinating between many threads. 

One way to think of this is with https://en.wikipedia.org/wiki/Amdahl%27s_law. Where the amount of speedup in a system by adding a new resource is directly limited by the amount of coordination needed between resources. When a request is placed on the Puma queue, other threads all coordinate via mutexes and conditiona variables in-order to do book-keeping and only pull of the next  request in line. MRI also has a GIL/GVL, in addition to these locks, versus JRuby does not have a GIL, each thread can independently  run Ruby code. It seems that decreasing pressure on the locking mechanisms is much more important for JRuby than for Cruby (MRI).

A key fact in the prior "longtail" reproduction is that it only occurs when all Puma threads are busy, that is when Puma is overloaded. That is because if a thread is busy working on several requests in a row and a new request comes in, but can be picked up by another thread ... that request does not have to wait and can be served fairly.

## This change

This PR reintroduces the prior "fast inline" behavior to allow a thread to be monopolized, but with a key difference. This will only be allowed when Puma detects that there are idle threads waiting for a new request. If all threads are busy, it will put the next request on the connection in the queue for proper ordering. This approach preserves the Puma 7 "fair" ordering in addition to reintroducing the Puma 6 "fast inline" optimization.

Puma 6 had a setting `max_fast_inline` that would auto close a connection after a number of requests on it were served (default was 10). That was needed to prevent a DoS as otherwise an attacker could saturate all threads with keepalive connections and then prevent any new requests from being served. This change does NOT reintroduce this concept. Connections will be closed after they've reached the maximum keepalive number of requests (default is 999). As long as the system has excess capacity (spare threads), it is not at risk of DoS via this vector. As soon as all threads are saturated, it will not close the connection (as before) but instead place it in queue.

## Reproduction

Performance comparison using this script https://gist.github.com/schneems/bc7ee417007d467a63abcf793793b989/b51629f86f6e05e2afc52041875033c6bffa5012

```
$ git co <branchname> && ./tmp/profile_puma.sh
```

Results:

```
v6.6.1    Requests/sec: 116,403.14
v7.04     Requests/sec:  20,968.15
My branch Requests/sec: 169,514.68
```

Higher is better. Not only does this meet the performance of Puma 6, but it beats it.

close #3788