Implement prototype for IO bound processor threads
Fix: https://github.com/puma/puma/issues/3777

This is only meant as a rought prototype to better illustrate
the idea. There's no extra tests to speak of, but the crux of
the logic is here.

Workers can now be marked as "IO bound".
This state last until the end of the current request.

At the end of the request, the pool decide whether the "IO bound" thread
rejoin the normal population, or is abandonned.

Generally speaking, an "IO bound" thread no longer counts towards
the `max_threads` configuration. However, there is a `max_io_threads`
configuration, and any "IO bound" thread above that limit automatically
counts as a "normal" thread.

For example, imagine a config of `max_threads: 5, max_io_threads: 10`:

If there is 3 normal threads and 12 io threads, then puma won't spawn
any new thread, but the next deferred thread that complete will become "normal"
again so that we now have 4 normal threads and 11 deferred threads.

Now if there is 5 normal threads and 2 IO threads, then when the next
IO threads complete, it will exit.