Begin passing HTTP read buffer directly through parser
As mentioned in the deleted comment, the parser should be able to
operate directly against the incoming bytes, rather than making a
copy. This commit does so, parsing directly against the incoming
bytes and passing them to the leaf methods, which make their own
copies of the fields and values as appropriate.