Construct all parsed strings as shared
While most headers are either pre-allocated or cached as fstrings,
header values and query strings must be allocated new for each
request. By marking the incoming buffer as shared and allocating
other strings as shared views into that buffer's bytes, we can
avoid copying those bytes again for each string as is done in
CRuby.

Marking the original buffer as shared does mean the request logic
will be forced to create a new ByteList and byte[] if further data
is read, but this is no worse (and probably better) than makiong N
copied slices of the buffer while parsing the request.

Because of this copy-on-write behavior, the entirei incoming buffer
will be held in memory for the duration of the request, but most of
that data would be in memory until processed anyway, potentially
with a larger actual heap size than just a single large shared
buffer.