Perfect hashing and cache for field and CGI strings
This removes the linear search for matching field names and
replaces it with a perfect hash based on the standard HTTP/1.1
field names (plus some additions) and the supported CGI variables.

The array of pre-allocated strings is passed through the parser
from Http11.

The EnvKey enum moves to its own top-level class, and now contains
a gperf-generated perfect hash function to quickly find or reject
incoming fields.

Along with the EnvKey changes, the snake-upcasing of incoming
fields is done lazily, and never writes to the original buffer.
I have reverted a previous commit and restored the JRuby-specific
assertions related to this change, and I believe the direct
modification of the read buffer is a bug, albeit not a very
important one since it only affects error messages for malformed
HTTP requests.